requirements.txt
apache-airflow-providers-mongo
apache-airflow-providers-discord
pytest
pydantic
pydantic-mongo